Create Password checking
password = 12345
If the user enters the wrong password, the program will ask to enter the password again. If the user enters the wrong password 5 times, the program will terminate or close. If the user enters the right password after the wrong password is entered, the program will welcome the user and terminates.